# web/openapi.json
{
  "components": {
    "schemas": {
      "AgentConfig": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "default": "",
            "title": "Base Url",
            "type": "string"
          },
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "context_window": {
            "default": 1000000,
            "minimum": 0.0,
            "title": "Context Window",
            "type": "integer"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "use_responses": {
            "default": false,
            "title": "Use Responses",
            "type": "boolean"
          }
        },
        "title": "AgentConfig",
        "type": "object"
      },
...
      "AgentPoolConfig": {
        "additionalProperties": false,
        "properties": {
          "max_size": {
            "default": 256,
            "minimum": 1.0,
            "title": "Max Size",
            "type": "integer"
          },
          "sweep_interval_seconds": {
            "default": 60,
            "minimum": 1.0,
            "title": "Sweep Interval Seconds",
            "type": "integer"
          },
          "ttl_seconds": {
            "default": 1800,
            "minimum": 0.0,
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "title": "AgentPoolConfig",
        "type": "object"
      },
      "AgentRuntimeConfig": {
        "additionalProperties": false,
        "properties": {
          "context_budget_model_call_ratio": {
            "default": 0.8,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Budget Model Call Ratio",
            "type": "number"
          },
          "context_compression_enabled": {
            "default": true,
            "title": "Context Compression Enabled",
            "type": "boolean"
          },
          "context_compression_hard_stop_ratio": {
            "default": 0.98,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Hard Stop Ratio",
            "type": "number"
          },
          "context_compression_min_items": {
            "default": 12,
            "minimum": 1.0,
            "title": "Context Compression Min Items",
            "type": "integer"
          },
          "context_compression_preserve_recent_items": {
            "default": 20,
            "minimum": 1.0,
            "title": "Context Compression Preserve Recent Items",
            "type": "integer"
          },
          "context_compression_preserve_recent_ratio": {
            "default": 0.25,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Preserve Recent Ratio",
            "type": "number"
          },
          "context_compression_summary_max_tokens": {
            "default": 8000,
            "minimum": 512.0,
            "title": "Context Compression Summary Max Tokens",
            "type": "integer"
          },
          "context_compression_target_ratio": {
            "default": 0.2,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Target Ratio",
            "type": "number"
          },
          "context_compression_trigger_ratio": {
            "default": 0.9,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Trigger Ratio",
            "type": "number"
          },
          "main_max_turns": {
            "default": 1000,
            "minimum": 1.0,
            "title": "Main Max Turns",
            "type": "integer"
          },
          "model_stream_idle_timeout_seconds": {
            "default": 300,
            "minimum": 30.0,
            "title": "Model Stream Idle Timeout Seconds",
            "type": "integer"
          },
          "report_retention_seconds": {
            "default": 259200,
            "minimum": 0.0,
            "title": "Report Retention Seconds",
            "type": "integer"
          },
          "subordinate_max_turns": {
            "default": 1000,
            "minimum": 1.0,
            "title": "Subordinate Max Turns",
            "type": "integer"
          }
        },
        "title": "AgentRuntimeConfig",
        "type": "object"
      },
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_FetchProviderModelsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FetchProviderModelsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[FetchProviderModelsResponse]",
        "type": "object"
      },
      "CommonResponse_InstanceConfigSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstanceConfigSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[InstanceConfigSchema]",
        "type": "object"
      },
...
      "CommonResponse_UpdateInstanceConfigResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UpdateInstanceConfigResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[UpdateInstanceConfigResponse]",
        "type": "object"
      },
...
      "FetchProviderModelsRequest": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "default": "",
            "maxLength": 4096,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          }
        },
        "required": [
          "base_url"
        ],
        "title": "FetchProviderModelsRequest",
        "type": "object"
      },
      "FetchProviderModelsResponse": {
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "models"
        ],
        "title": "FetchProviderModelsResponse",
        "type": "object"
      },
...
      "InstanceConfigSchema": {
        "additionalProperties": false,
        "properties": {
          "agent_pool": {
            "$ref": "#/components/schemas/AgentPoolConfig"
          },
          "agent_runtime": {
            "$ref": "#/components/schemas/AgentRuntimeConfig"
          },
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentConfig"
            },
            "title": "Agents",
            "type": "object"
          },
          "lightrag": {
            "$ref": "#/components/schemas/LightRAGConfig"
          }
        },
        "title": "InstanceConfigSchema",
        "type": "object"
      },
...
      "LightRAGConfig": {
        "additionalProperties": false,
        "properties": {
          "chunk_matches": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Chunk Matches",
            "type": "integer"
          },
          "embedding_api": {
            "default": "https://api.openai.com/v1",
            "minLength": 1,
            "title": "Embedding Api",
            "type": "string"
          },
          "embedding_dim": {
            "default": 1536,
            "minimum": 1.0,
            "title": "Embedding Dim",
            "type": "integer"
          },
          "embedding_key": {
            "default": "",
            "title": "Embedding Key",
            "type": "string"
          },
          "embedding_model": {
            "default": "text-embedding-3-small",
            "minLength": 1,
            "title": "Embedding Model",
            "type": "string"
          },
          "graph_matches": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Graph Matches",
            "type": "integer"
          },
          "llm_api": {
            "default": "https://api.openai.com/v1",
            "minLength": 1,
            "title": "Llm Api",
            "type": "string"
          },
          "llm_key": {
            "default": "",
            "title": "Llm Key",
            "type": "string"
          },
          "llm_model": {
            "default": "gpt-5",
            "minLength": 1,
            "title": "Llm Model",
            "type": "string"
          }
        },
        "title": "LightRAGConfig",
        "type": "object"
      },
...
      "UpdateAgentConfigRequest": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "context_window": {
            "minimum": 0.0,
            "title": "Context Window",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "title": "Name",
            "type": "string"
          },
          "use_responses": {
            "title": "Use Responses",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "description",
          "base_url",
          "api_key",
          "model",
          "use_responses",
          "context_window"
        ],
        "title": "UpdateAgentConfigRequest",
        "type": "object"
      },
...
      "UpdateInstanceConfigRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_pool": {
            "$ref": "#/components/schemas/AgentPoolConfig"
          },
          "agent_runtime": {
            "$ref": "#/components/schemas/AgentRuntimeConfig"
          },
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/UpdateAgentConfigRequest"
            },
            "title": "Agents",
            "type": "object"
          },
          "lightrag": {
            "$ref": "#/components/schemas/LightRAGConfig"
          }
        },
        "required": [
          "agents",
          "agent_pool",
          "agent_runtime",
          "lightrag"
        ],
        "title": "UpdateInstanceConfigRequest",
        "type": "object"
      },
      "UpdateInstanceConfigResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/InstanceConfigSchema"
          },
          "restarted": {
            "title": "Restarted",
            "type": "boolean"
          }
        },
        "required": [
          "config",
          "restarted"
        ],
        "title": "UpdateInstanceConfigResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "ZJ - Multi-Agent Cyber Operations Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/system-config/instance": {
      "get": {
        "operationId": "get_instance_config_handler_api_system_config_instance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_InstanceConfigSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Get Instance Config Handler",
        "tags": [
          "system-config"
        ]
      },
      "patch": {
        "operationId": "update_instance_config_handler_api_system_config_instance_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_UpdateInstanceConfigResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Instance Config Handler",
        "tags": [
          "system-config"
        ]
      }
    },
    "/api/system-config/models": {
      "post": {
        "operationId": "fetch_provider_models_handler_api_system_config_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchProviderModelsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_FetchProviderModelsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Fetch Provider Models Handler",
        "tags": [
          "system-config"
        ]
      }
    },
...
  }
}